Aryu/requirements.txt
Django
djangorestframework
channels
channels-redis
django-cors-headers
django-allauth
dj-rest-auth
django-redis
django-axes
django-crontab
psycopg2-binary
stripe
razorpay
twilio
paypalrestsdk
reportlab
pytesseract
pyclamd
Pillow
qrcode
django-countries
django-filter
captcha
holidays
num2words
pytz

